Update RoCC add test to use int64
This more closely matches what VCODE & NESL use, which only use 64-bit
signed integers and 64-bit double-precision floating-point numbers,
along with booleans and special integers for file operations.

+++ test/src/rocc_add.c
@@ -1,7 +1,8 @@
 #include <rocc.h>
+#include <stdint.h>
 
 int main() {
-    int a,b,c;
+    int64_t a,b,c;
     a = 1;
     b = 2;
     ROCC_INSTRUCTION_DSS(0, c, &a, &b, 0);
